Change to default button
