+++ packages/experiments/src/components/Button/DefaultButton/DefaultButton.tsx
@@ -1,0 +1,10 @@
+import { styled } from '../../../Utilities';
+import { SplitButtonBase } from '../_base/SplitButton.base';
+
+
+import { getStyles } from './DefaultButton.styles';
+
+export const DefaultSplitButton = styled(
+  SplitButtonBase,
+  getStyles
+);
